iter18 fix #2
